feat : Enrollment 추상화 삭제 (피드백 반영)
